bot/requirements.txt
aiogram==3.4.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reportlab==4.2.5

